Added a wave system, modified timer
Added a game manger to keep track of waves and time. The timer and wave texts are both set through the game manager. The game manager also has an update wave function to handle what happens every time a new wave starts - we can handle difficulty increases and stuff in this method.